# [Merged by Bors] - Use `SmallVec` for `TreeHash` packed encoding