--- a/consensus/types/src/slot_epoch.rs
+++ b/consensus/types/src/slot_epoch.rs
@@ -16,7 +16,7 @@ use crate::{ChainSpec, SignedRoot};
 use rand::RngCore;
 use safe_arith::{ArithError, SafeArith};
 use serde_derive::{Deserialize, Serialize};
-use ssz::{ssz_encode, Decode, DecodeError, Encode};
+use ssz::{Decode, DecodeError, Encode};
 use std::fmt;
 use std::hash::Hash;
 use std::iter::Iterator;