--- a/consensus/types/src/slot_epoch_macros.rs
+++ b/consensus/types/src/slot_epoch_macros.rs
@@ -291,7 +291,7 @@ macro_rules! impl_ssz {
             }
 
             fn tree_hash_packed_encoding(&self) -> tree_hash::PackedEncoding {
-                tree_hash::PackedEncoding::from_slice(&self.0.to_le_bytes())
+                self.0.tree_hash_packed_encoding()
             }
 
             fn tree_hash_packing_factor() -> usize {